start of Binary Tree implementation
Now i started to implement the Binary Tree, for now it has two methods, add a new element in the tree and show itself. I decided to use recursion with these methods because i thought it's more easy to implement like this, trees in general it's better to implement using recursion my professor said and trying to implement it now i agree with her.